Removed the (to be deprecated) iframe from the movieplayer. Placed a div instead.
The menu is now identified by an id. A class attribute will decide whether to use a vertical or horizontal layout.

Javascript added to dynamically change the size of the window, depending on the high or low quality video. It also changes the class
of the menu, meaning the layout of the menu can be changed dynamically as well. Right now, the menu is disabled with a high quality video.